Pull in applicable changes from the toolkit branch:
r4439: 
Bump the version # on the ls cache daemon

r4438: 
Archive::Tar bails out if '.' or '..' are in the tarball. Generally, this is not the behavior we'd like, so remove them from the files to extract

r4379: 
Bump the release information on the ls cache daemon to minor release 6


r4376: 
Get the LS cache daemon working with an older version of Archive::Tar
(available as an RPM so we won't have to package it ourselves).

r4258: 
Change the log rotation to daily for the pS services

r4145: 
RedHat's shutdown scripts require a file in "/var/lock/subsys" before they will
run any init scripts to shut down the service (it just silently doesn't call
them if it's missing).  Update the various init scripts to place that file
there so that the init scripts will actually get called on shutdown.



git-svn-id: https://svn.internet2.edu/svn/perfSONAR-PS/trunk@4441 6dbdcc9b-cb27-0410-b2e0-cf3f5b33df3b